@@ -8,9 +8,7 @@
 import Foundation
 import Network
 
-typealias SocketResponseCompletion = @Sendable (Result<SocketResponse, Error>) -> Void
-
-public final class SocketWrapper: Sendable {
+public final actor AsyncSocketWrapper: Sendable {
     private let requestMarker: [UInt8]
     private let host: NWEndpoint.Host
     private let port: NWEndpoint.Port
@@ -20,143 +18,134 @@ public final class SocketWrapper: Sendable {
     private nonisolated(unsafe) var startingTime: TimeInterval = 0
     private nonisolated(unsafe) var timer: Timer?
     private let queue = DispatchQueue(label: "com.gameServerQueryLibrary.socketQueue")
-    private nonisolated(unsafe) var completionHandler: SocketResponseCompletion?
-    private let dispatchQueue = DispatchQueue(label: "com.gsql.socket-queue")
-
-    deinit {
-        cleanup()
-    }
+    private var continuation: CheckedContinuation<SocketResponse, Error>?
     
-    init(requestMarker: [UInt8], host: NWEndpoint.Host, port: NWEndpoint.Port) {
+    private var requestInProgress = false
+    
+    public init(requestMarker: [UInt8], host: NWEndpoint.Host, port: NWEndpoint.Port) {
         self.requestMarker = requestMarker
         self.host = host
         self.port = port
         connection = NWConnection(host: host, port: port, using: .udp)
-        observeConnectionStateUpdates()
         connection.start(queue: queue)
+        Task {
+            await observeConnectionStateUpdates()
+        }
     }
     
-    private func observeConnectionStateUpdates() {
-        connection.stateUpdateHandler = { [weak self] (newState) in
-            self?.dispatchQueue.async { [weak self] in
-                guard let self else {
-                    return
+    public func observeConnectionStateUpdates() async {
+        connection.stateUpdateHandler = { [weak self] newState in
+            Task {
+                await self?.handleNewState(newState)
+            }
+        }
+    }
+    
+    public func sendRequest() async throws -> SocketResponse {
+        guard !requestInProgress else {
+            throw SocketError.requestAlreadyInProgress
+        }
+        requestInProgress = true
+        return try await withCheckedThrowingContinuation { continuation in
+            self.continuation = continuation
+            Task {
+                if connection.state == .ready {
+                    let data = Data(requestMarker)
+                    await sendRequest(data)
                 }
-                switch (newState) {
-                case .ready:
-                    if sendRequestInitiated {
-                        let data = Data(requestMarker)
-                        sendRequest(data)
-                    }
-                case .failed(let error):
-                    self.completionHandler?(.failure(error))
-                default:
-                    break
+            }
+        }
+    }
+    
+    private func finish(with response: SocketResponse) async {
+        continuation?.resume(returning: response)
+    }
+    
+    private func finish(with error: Error) async {
+        continuation?.resume(throwing: error)
+    }
+    
+    private func cleanup() async {
+        continuation = nil
+        await invalidateTimer()
+        connection.cancel()
+    }
+    
+    private func handleNewState(_ state: NWConnection.State) async {
+        switch state {
+            case .ready:
+                if requestInProgress {
+                    let data = Data(requestMarker)
+                    await sendRequest(data)
                 }
-            }
+            case .failed(let error):
+                await self.finish(with: error)
+            default:
+                break
         }
     }
     
-    func sendRequest(_ completionHandler: @escaping SocketResponseCompletion) {
-        dispatchQueue.async { [weak self] in
-            guard let self else {
-                return
-            }
-            sendRequestInitiated = true
-            self.completionHandler = completionHandler
-            
-            if connection.state == .ready {
-                let data = Data(requestMarker)
-                sendRequest(data)
-            }
-        }
-    }
-    
-    func startTimer() {
-        DispatchQueue.main.async { [weak self] in
-            guard let self else {
-                return
-            }
-            self.timer = Timer.scheduledTimer(timeInterval: 0.5, target: self, selector: #selector(self.didNotReceiveResponseInTime), userInfo: nil, repeats: false)
-        }
-    }
-    
-    func cancel() {
-        cleanup()
-    }
-
-    private func sendRequest(_ content: Data) {
+    private func sendRequest(_ content: Data) async {
         connection.send(content: content, completion: NWConnection.SendCompletion.contentProcessed(({ [weak self] error in
-            self?.dispatchQueue.async { [weak self] in
-                if let error {
-                    self?.completionHandler?(.failure(error))
-                } else {
-                    self?.startingTime = Date.timeIntervalSinceReferenceDate
+            if let error {
+                Task { [weak self] in
+                    await self?.finish(with: error)
                 }
+            } else {
+                self?.startingTime = Date.timeIntervalSinceReferenceDate
             }
         })))
         
-        startTimer()
+        await startTimer()
         
         listenForDatagrams()
     }
-
+    
     private func listenForDatagrams() {
-        connection.receive(minimumIncompleteLength: 1, maximumLength: 8_192) { [weak self] (data, context, isComplete, error) in
-            self?.dispatchQueue.async { [weak self] in
+        connection.receive(minimumIncompleteLength: 1, maximumLength: 8_192) { (data, context, isComplete, error) in
+            Task { [weak self] in
                 guard let self else {
                     return
                 }
                 if isComplete {
-                    self.invalidateTimer()
+                    await self.invalidateTimer()
                     if let data {
                         self.data.append(data)
                     }
                     let runningTime = ((Date.timeIntervalSinceReferenceDate - self.startingTime) * 1000).rounded()
                     let response = SocketResponse(data: self.data, runningTime: Int(runningTime))
-                    self.completionHandler?(.success(response))
+                    await finish(with: response)
                     return
                 }
                 if let error {
                     NLog.error(error)
-                    self.completionHandler?(.failure(error))
+                    await finish(with: error)
                     return
                 } else if let data {
                     self.data.append(data)
-                    self.listenForDatagrams()
+                    await self.listenForDatagrams()
                 }
             }
         }
     }
     
-    @objc private func didNotReceiveResponseInTime(_ timer: Timer) {
-        dispatchQueue.async { [weak self] in
-            guard let self else {
-                return
-            }
-            completionHandler?(.failure(SocketError.timeout(host.debugDescription, port.rawValue)))
-            cleanup()
+    private func startTimer() async {
+        await MainActor.run {
+            timer = Timer.scheduledTimer(timeInterval: 0.5, target: self, selector: #selector(self.didNotReceiveResponseInTime), userInfo: nil, repeats: false)
         }
     }
     
-    private func invalidateTimer() {
-        DispatchQueue.main.async { [weak self] in
-            guard let self else {
-                return
-            }
-            self.timer?.invalidate()
-            self.timer = nil
+    @objc private func didNotReceiveResponseInTime() async {
+        Task {
+            await finish(with: SocketError.timeout(host.debugDescription, port.rawValue))
+            await cleanup()
         }
     }
     
-    private func cleanup() {
-        dispatchQueue.async { [weak self] in
-            guard let self else {
-                return
-            }
-            invalidateTimer()
-            completionHandler = nil
-            connection.cancel()
+    private func invalidateTimer() async {
+        await MainActor.run {
+            timer?.invalidate()
+            timer = nil
         }
     }
 }
