@@ -10,7 +10,7 @@ import Foundation
 
 public protocol Coordinator {
     func getServersList(ip: String, port: String) async throws -> [Server]
-    func fetchServersInfo(for servers: [Server], waitTimeInMilliseconds: TimeInterval) -> AsyncStream<Server>
+    func fetchServersInfo(for servers: [Server], waitTimeInMilliseconds: TimeInterval) -> AsyncThrowingStream<Server, Error>
     func updateServerInfo(_ server: Server) async throws -> Server
     func updateServerStatus(_ server: Server) async throws -> Server
 }
